Rename the discharge-diceased-patient-modal folder
