+++ ui/src/app/modules/mortuary/pages/mortuary-dashboard/mortuary-dashboard.component.ts
@@ -27,7 +27,7 @@ import { getCustomOpenMRSFormsByIds } from "src/app/store/selectors/form.selecto
 import { ICARE_CONFIG } from "src/app/shared/resources/config";
 import { EncountersService } from "src/app/shared/services/encounters.service";
 import { DischargePatientModalComponent } from "src/app/shared/components/discharge-patient-modal/discharge-patient-modal.component";
-import { DischargeDeceasedPatientModalComponent } from "src/app/shared/components/discharge-deceased-patient-modal /discharge-deceased-patient-modal.component";
+import { DischargeDeceasedPatientModalComponent } from "src/app/shared/components/discharge-deceased-patient-modal/discharge-deceased-patient-modal.component";
 
 @Component({
   selector: "app-mortuary-dashboard",
@@ -155,16 +155,14 @@ export class MortuaryDashboardComponent implements OnInit {
     });
   }
 
-  onDischarge(event: Event,visit, patient, provider): void {
-  
+  onDischarge(event: Event, visit, patient, provider): void {
     event.stopPropagation();
     this.dialog.open(DischargeDeceasedPatientModalComponent, {
       minWidth: "30%",
       data: {
-         visit,
-         provider,
-         patient
-        
+        visit,
+        provider,
+        patient,
       },
     });
   }
