+++ ui/src/app/shared/components/discharge-deceased-patient-modal/discharge-deceased-patient-modal.component.ts
