+++ ui/src/app/shared/components/index.ts
@@ -124,7 +124,7 @@ import { SharedRenderReportDashboardComponent } from "./shared-render-report-das
 import { SharedDrugDosageInterpretationComponent } from "./shared-drug-dosage-interpretation/shared-drug-dosage-interpretation.component";
 import { SharedLocationBasedLogoComponent } from "./shared-location-based-logo/shared-location-based-logo.component";
 import { SharedPatientDischargeDetailsComponent } from "./shared-patient-discharge-details/shared-patient-discharge-details.component";
-import { DischargeDeceasedPatientModalComponent } from "./discharge-deceased-patient-modal /discharge-deceased-patient-modal.component";
+import { DischargeDeceasedPatientModalComponent } from "./discharge-deceased-patient-modal/discharge-deceased-patient-modal.component";
 import { SharedNextOfKinsFormComponent } from "./shared-next-of-kins-form/shared-next-of-kins-form.component";
 import { SharedNextOfKinsFormDataComponent } from "./shared-next-of-kins-form-data/shared-next-of-kins-form-data.component";
